container/requirements.txt
# PyPI library used in all projects
rdetoolkit==1.3.4
# PyPI libraries to be customized and installed in each project
matplotlib==3.8.0
pdf2image==1.17.0
python-pptx==0.6.23
scipy==1.13.1
# Only required when attaching to Docker for debugging.
# debugpy